Release TEAMNADO-4772 to prod stable (#175)
* TEAMNADO-4641: Activation Key Details - Save additional repos (#160)

* Add functionality to 'Add Repositories' table and model to allow repos to be added to activation key

* Fixed linting issues

* Fixed more linting errors

* Attempting to conform to linting rules

* Fixed final linting issues

* Made changes to resolve discussions

* Updated test cases to reflect previous changes

* Updated test

* TEAMNADO-4772 - Abstract common logic (#172)

* gated rbac button

* Fix linting errors

* Rename EditAdditionalRepositoriesButton to AddAdditionalRepositoriesButton

---------
