+++ src/Components/WriteOnlyButton/index.js
@@ -1,0 +1,3 @@
+import WriteOnlyButton from './WriteOnlyButton';
+
+export { WriteOnlyButton };
